adblock: reset error to NULL if list regexp could not be parsed
Reset (freed) error variable to NULL after a regexp parsing error
from the whitelist/blacklist file. It's reused in the next loop
iteration and if we don't reset it, we get critical warnings when
reading further regexp lines, and then crashes when trying to
print or free the already-freed error.

https://bugzilla.gnome.org/show_bug.cgi?id=595094